+++ src/App.js
@@ -1,35 +1,19 @@
-import { useState, Fragment, useEffect } from "react";
+import { useState, Fragment, useContext } from "react";
 
 import logo from "./logo.svg";
 import "./App.css";
 import MainForm from "./components/MainForm";
 import ShiftList from "./components/ShiftsList";
-import Login from "./components/Login/Login";
 import MainHeader from "./components/UI/MainHeader";
 import AuthContext from "./store/auth-context";
+import Login from "./components/Login/Login";
 
 const App = (props) => {
+  const AuthCtx = useContext(AuthContext);
+
   const [list, setList] = useState([]);
-  const [isLoggedIn, setIsLoggedIn] = useState(false);
+
   const [amount, setAmount] = useState(``);
-
-  useEffect(() => {
-    const storedUserLoggedInInfo = localStorage.getItem("isLoggedIn");
-
-    if (storedUserLoggedInInfo === "1") {
-      setIsLoggedIn(true);
-    }
-  });
-
-  const loginHandler = (email, password) => {
-    localStorage.setItem("isLoggedIn", "1");
-    setIsLoggedIn(true);
-  };
-
-  const logoutHandler = () => {
-    localStorage.removeItem("isLoggedIn");
-    setIsLoggedIn(false);
-  };
 
   const addEntryHandler = (date, from, to, amount, shiftTime) => {
     setList((prevShiftsList) => {
@@ -45,21 +29,16 @@ const App = (props) => {
   };
 
   return (
-    <AuthContext.Provider
-      value={{
-        isLoggedIn: isLoggedIn,
-        onLogout: logoutHandler,
-      }}
-    >
-      {!isLoggedIn && (
+    <Fragment>
+      {!AuthCtx.isLoggedIn && (
         <Fragment>
-          <MainHeader onLogout={logoutHandler} />
-          <Login onLogin={loginHandler} />
+          <MainHeader onLogout={AuthCtx.logoutHandler} />
+          <Login onLogin={AuthCtx.loginHandler} />
         </Fragment>
       )}
-      {isLoggedIn && (
+      {AuthCtx.isLoggedIn && (
         <Fragment>
-          <MainHeader onLogout={logoutHandler} />
+          <MainHeader onLogout={AuthCtx.logoutHandler} />
           <MainForm
             addEntryMainForm={addEntryHandler}
             addAmountMainForm={addAmountHandler}
@@ -67,7 +46,7 @@ const App = (props) => {
           <ShiftList shifts={list} />
         </Fragment>
       )}
-    </AuthContext.Provider>
+    </Fragment>
   );
 };
